requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
Click
easydict
ipdb
matplotlib
ninja
pykeops
numpy==1.24.2
open3d==0.17.0
pytorch_lightning==1.6
PyYAML
scikit_learn
scipy
torch==1.12.1+cu113
tqdm
hydra-core
more_itertools
torchtyping
cmapy
networkx
fpsample
natsort
diskcache
fpsample
diffusers
torchtyping 
cmapy 
diskcache
torch_geometric
networkx
pybind11 
setuptools
plyfile